Refactor `package sync` implementation to be generic
Prepare implementation to also work for `component sync`.
